Color and Rainbow Utilities
Better.

+++ summon.py
@@ -9,22 +9,27 @@ def summon(inp, chan=None, conn=None):
 	if chan == "#eisenstein":
 		if inp == "!all":
 			open("summonlist.txt", 'a').close()
+			lockfile(chan)
 			conn.cmd("WHO", [chan])
 		else:
 			return "Calling " + inp + "."
 			
 @hook.event('352')
 def who_summon(inp, conn=None, chan="#eisenstein"):
-	user = str(inp[5]);
-	store(user + ";")
+	channel = str(inp[1])
+	if channel == "#eisenstein":
+		user = str(inp[5]);
+		store(user + ";")
 	
 @hook.event('315')
 def end_summon(inp, conn=None, chan="#eisenstein"):
-	usrlist = retreive()
-	del usrlist[-1]
-	prettylist = ', '.join(usrlist)
-	announce(prettylist, conn, "#eisenstein");
-	os.remove("summonlist.txt")
+	if os.path.exists("summon.lock"):
+		usrlist = retreive()
+		del usrlist[-1]
+		prettylist = ', '.join(usrlist)
+		announce(prettylist, conn, "#eisenstein");
+		os.remove("summonlist.txt")
+		os.remove("summon.lock")
 
 def store(data):
 	try:
@@ -49,3 +54,14 @@ def retreive():
 		string = "Could not read file."
 	arguments = args.split(";")
 	return arguments;
+	
+def lockfile(data):
+	try:
+		#Lockfile for tracking
+		summonlock = open("summon.lock", "a")
+		try:
+			summonlock.write(data)
+		finally:
+			summonlock.close()
+	except IOError:
+		pass
